requirements.txt
PyYAML~=6.0.1
pygame~=2.6.0
PyDirectInput~=1.0.4
chardet~=5.2.0
pynput~=1.7.7
PySide6~=6.7.2
PySide6-Fluent-Widgets[full]~=1.6.5
numpy~=2.1.1
pillow~=10.4.0
requests~=2.32.3
pydantic~=2.9.2